@@ -29,16 +29,16 @@
     "babel-preset-react": "^6.11.1",
     "babel-preset-stage-0": "^6.5.0",
     "chai": "^3.5.0",
-    "eslint": "^3.2.2",
-    "eslint-config-airbnb": "^12.0.0",
-    "eslint-plugin-import": "^1.13.0",
-    "eslint-plugin-jsx-a11y": "^2.1.0",
-    "eslint-plugin-react": "^6.0.0",
+    "eslint": "^3.13.0",
+    "eslint-config-airbnb": "^14.0.0",
+    "eslint-plugin-import": "^2.2.0",
+    "eslint-plugin-jsx-a11y": "^3.0.2",
+    "eslint-plugin-react": "^6.9.0",
     "gulp": "^3.9.1",
     "gulp-eslint": "^3.0.1",
     "gulp-header": "^1.8.8",
-    "gulp-mocha-phantomjs": "^0.11.0",
-    "gulp-sass": "^2.3.2",
+    "gulp-mocha-phantomjs": "^0.12.0",
+    "gulp-sass": "^3.0.0",
     "gulp-scss-lint": "^0.4.0",
     "http-server": "^0.9.0",
     "mocha": "^3.0.2",
